+++ src/components/Hidden/tests/Hidden.stories.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Example } from "utilities/storybook";
 import Hidden from "components/Hidden";